tipc: initialize option list when creating cmd struct
The option list was initialized in the constructor function for the
command. Real-world testing suggests this might segfault on some systems
as the constructor functions are not called in the correct order within
the compilation unit.

Initialize the option list when creating the cmd struct instead. This
should allow for out-of-order creation.

Signed-off-by: Andreas Bofjall <andreas@gazonk.org>
Signed-off-by: Richard Alpe <richard.alpe@ericsson.com>