tipc: redesign socket list
To accommodate kernel space changes we no longer reassemble
publications for the same socket. We now ask for a list of sockets
with TIPC_NL_SOCK_GET and if a socket has publications
(TIPC_NLA_SOCK_HAS_PUBL) we ask for them using TIPC_NL_PUBL_GET.

This is fairly slow for the worst case which is a high amount of
published sockets with only one or a few publications.