Added new subtests to multicast test.
The multicast test in this test suite has so far been totally inadequate.
Multicast is basically a separate stack, and must be tested in a similar
way we do with unicast.

In this commit we add three new subtests:

- Sending big bursts of small messages, hence testing that bundling
  functionalit works on the braodcast link.
- Sending large bursts of 1k messages, implying a test of broadcast
  link congestion and user wakeup.
- Sending a maximum burst (40) of 66000 byte messages, hence testing
  fragmentation/reassembly and link congestion.

In addition, all multicast messages are now subject to integrity test,
using the functions provided by the test suite.