tipcutils: added sequence numbers to multicast test messages
Sometimes a message may show up prematurely in a socket, while
the server is still reading the previous message from the socket
set. This causes confusion, since the server exits the reception
loop too early, thinking it has received messages on all expected
sockets.
We now peek the message first, and if the sequence number is not
the expected one, we ignore it, leaving it to be caught at next
reception.

received the